Moved LoadImageFile And LoadLabel to SharedLIb
