+++ Samples/SampleSharedLib/SampleSharedLib/pch.h
@@ -1,6 +1,12 @@
 #ifndef PCH_H
 #define PCH_H
 
+#include <winrt/Windows.Graphics.Imaging.h>
+#include <winrt/Windows.Storage.h>
+#include "winrt/Windows.Storage.Streams.h"
+
+#include <Windows.h>
+
 #include <string>
 #include <fstream>
